Backend/requirements.txt
# ── Core (required) ──────────────────────────────────────────────────────────
Django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.11.11
whitenoise==6.8.2

# ── ML (required for AI features) ────────────────────────────────────────────
numpy>=2.0
scikit-learn>=1.5
joblib>=1.3

# ── Production extras (optional) ─────────────────────────────────────────────
# psycopg2-binary==2.9.9
# celery==5.4.0
# redis==5.2.1
